+++ lib/src/vascular/inbox.pb.dart
@@ -3,12 +3,15 @@
 //  source: inbox.proto
 //
 // @dart = 2.12
-// ignore_for_file: annotate_overrides,camel_case_types,constant_identifier_names,directives_ordering,library_prefixes,non_constant_identifier_names,prefer_final_fields,return_of_invalid_type,unnecessary_const,unnecessary_import,unnecessary_this,unused_import,unused_shown_name
+// ignore_for_file: annotate_overrides,camel_case_types,unnecessary_const,non_constant_identifier_names,library_prefixes,unused_import,unused_shown_name,return_of_invalid_type,unnecessary_this,prefer_final_fields
 
 import 'dart:core' as $core;
-import 'dart:convert' as $convert;
 
 import 'package:protobuf/protobuf.dart' as $pb;
+
+import 'message.pb.dart' as $0;
+
+import 'message.pbenum.dart' as $0;
 
 class GetInboxMessagesRequest extends $pb.GeneratedMessage {
   static final $pb.BuilderInfo _i = $pb.BuilderInfo(const $core.bool.fromEnvironment('protobuf.omit_message_names') ? '' : 'GetInboxMessagesRequest', package: const $pb.PackageName(const $core.bool.fromEnvironment('protobuf.omit_message_names') ? '' : 'vascularinbox.inbox'), createEmptyInstance: create)
@@ -16,6 +19,7 @@ class GetInboxMessagesRequest extends $pb.GeneratedMessage {
     ..aOS(2, const $core.bool.fromEnvironment('protobuf.omit_field_names') ? '' : 'userId')
     ..aOS(3, const $core.bool.fromEnvironment('protobuf.omit_field_names') ? '' : 'hwId')
     ..aOM<Next>(4, const $core.bool.fromEnvironment('protobuf.omit_field_names') ? '' : 'next', subBuilder: Next.create)
+    ..pc<$0.Language>(5, const $core.bool.fromEnvironment('protobuf.omit_field_names') ? '' : 'langauges', $pb.PbFieldType.PE, valueOf: $0.Language.valueOf, enumValues: $0.Language.values)
     ..hasRequiredFields = false
   ;
 
@@ -25,6 +29,7 @@ class GetInboxMessagesRequest extends $pb.GeneratedMessage {
     $core.String? userId,
     $core.String? hwId,
     Next? next,
+    $core.Iterable<$0.Language>? langauges,
   }) {
     final _result = create();
     if (appKey != null) {
@@ -39,10 +44,13 @@ class GetInboxMessagesRequest extends $pb.GeneratedMessage {
     if (next != null) {
       _result.next = next;
     }
+    if (langauges != null) {
+      _result.langauges.addAll(langauges);
+    }
     return _result;
   }
   factory GetInboxMessagesRequest.fromBuffer($core.List<$core.int> i, [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) => create()..mergeFromBuffer(i, r);
-  factory GetInboxMessagesRequest.fromJson($core.String json) => GetInboxMessagesRequest.create()..mergeFromProto3Json($convert.jsonDecode(json));
+  factory GetInboxMessagesRequest.fromJson($core.String i, [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) => create()..mergeFromJson(i, r);
   @$core.Deprecated(
   'Using this can add significant overhead to your binary. '
   'Use [GeneratedMessageGenericExtensions.deepCopy] instead. '
@@ -99,6 +107,9 @@ class GetInboxMessagesRequest extends $pb.GeneratedMessage {
   void clearNext() => clearField(4);
   @$pb.TagNumber(4)
   Next ensureNext() => $_ensure(3);
+
+  @$pb.TagNumber(5)
+  $core.List<$0.Language> get langauges => $_getList(4);
 }
 
 class CreateInboxMessagesRequest extends $pb.GeneratedMessage {
@@ -138,7 +149,7 @@ class CreateInboxMessagesRequest extends $pb.GeneratedMessage {
     return _result;
   }
   factory CreateInboxMessagesRequest.fromBuffer($core.List<$core.int> i, [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) => create()..mergeFromBuffer(i, r);
-  factory CreateInboxMessagesRequest.fromJson($core.String json) => CreateInboxMessagesRequest.create()..mergeFromProto3Json($convert.jsonDecode(json));
+  factory CreateInboxMessagesRequest.fromJson($core.String i, [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) => create()..mergeFromJson(i, r);
   @$core.Deprecated(
   'Using this can add significant overhead to your binary. '
   'Use [GeneratedMessageGenericExtensions.deepCopy] instead. '
@@ -206,7 +217,7 @@ class CreateInboxMessagesRequest extends $pb.GeneratedMessage {
 
 class GetInboxMessagesReply extends $pb.GeneratedMessage {
   static final $pb.BuilderInfo _i = $pb.BuilderInfo(const $core.bool.fromEnvironment('protobuf.omit_message_names') ? '' : 'GetInboxMessagesReply', package: const $pb.PackageName(const $core.bool.fromEnvironment('protobuf.omit_message_names') ? '' : 'vascularinbox.inbox'), createEmptyInstance: create)
-    ..pc<InboxMessage>(1, const $core.bool.fromEnvironment('protobuf.omit_field_names') ? '' : 'messages', $pb.PbFieldType.PM, subBuilder: InboxMessage.create)
+    ..pc<$0.InboxMessage>(1, const $core.bool.fromEnvironment('protobuf.omit_field_names') ? '' : 'messages', $pb.PbFieldType.PM, subBuilder: $0.InboxMessage.create)
     ..pPS(2, const $core.bool.fromEnvironment('protobuf.omit_field_names') ? '' : 'newMessagesIds')
     ..pPS(3, const $core.bool.fromEnvironment('protobuf.omit_field_names') ? '' : 'readMessagesIds')
     ..aOM<Next>(4, const $core.bool.fromEnvironment('protobuf.omit_field_names') ? '' : 'next', subBuilder: Next.create)
@@ -216,7 +227,7 @@ class GetInboxMessagesReply extends $pb.GeneratedMessage {
 
   GetInboxMessagesReply._() : super();
   factory GetInboxMessagesReply({
-    $core.Iterable<InboxMessage>? messages,
+    $core.Iterable<$0.InboxMessage>? messages,
     $core.Iterable<$core.String>? newMessagesIds,
     $core.Iterable<$core.String>? readMessagesIds,
     Next? next,
@@ -241,7 +252,7 @@ class GetInboxMessagesReply extends $pb.GeneratedMessage {
     return _result;
   }
   factory GetInboxMessagesReply.fromBuffer($core.List<$core.int> i, [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) => create()..mergeFromBuffer(i, r);
-  factory GetInboxMessagesReply.fromJson($core.String json) => GetInboxMessagesReply.create()..mergeFromProto3Json($convert.jsonDecode(json));
+  factory GetInboxMessagesReply.fromJson($core.String i, [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) => create()..mergeFromJson(i, r);
   @$core.Deprecated(
   'Using this can add significant overhead to your binary. '
   'Use [GeneratedMessageGenericExtensions.deepCopy] instead. '
@@ -262,7 +273,7 @@ class GetInboxMessagesReply extends $pb.GeneratedMessage {
   static GetInboxMessagesReply? _defaultInstance;
 
   @$pb.TagNumber(1)
-  $core.List<InboxMessage> get messages => $_getList(0);
+  $core.List<$0.InboxMessage> get messages => $_getList(0);
 
   @$pb.TagNumber(2)
   $core.List<$core.String> get newMessagesIds => $_getList(1);
@@ -308,7 +319,7 @@ class CreateInboxMessagesReply extends $pb.GeneratedMessage {
     return _result;
   }
   factory CreateInboxMessagesReply.fromBuffer($core.List<$core.int> i, [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) => create()..mergeFromBuffer(i, r);
-  factory CreateInboxMessagesReply.fromJson($core.String json) => CreateInboxMessagesReply.create()..mergeFromProto3Json($convert.jsonDecode(json));
+  factory CreateInboxMessagesReply.fromJson($core.String i, [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) => create()..mergeFromJson(i, r);
   @$core.Deprecated(
   'Using this can add significant overhead to your binary. '
   'Use [GeneratedMessageGenericExtensions.deepCopy] instead. '
@@ -338,123 +349,6 @@ class CreateInboxMessagesReply extends $pb.GeneratedMessage {
   void clearInboxId() => clearField(1);
 }
 
-class InboxMessage extends $pb.GeneratedMessage {
-  static final $pb.BuilderInfo _i = $pb.BuilderInfo(const $core.bool.fromEnvironment('protobuf.omit_message_names') ? '' : 'InboxMessage', package: const $pb.PackageName(const $core.bool.fromEnvironment('protobuf.omit_message_names') ? '' : 'vascularinbox.inbox'), createEmptyInstance: create)
-    ..aOS(1, const $core.bool.fromEnvironment('protobuf.omit_field_names') ? '' : 'uuid')
-    ..a<$core.int>(2, const $core.bool.fromEnvironment('protobuf.omit_field_names') ? '' : 'status', $pb.PbFieldType.O3)
-    ..aOS(3, const $core.bool.fromEnvironment('protobuf.omit_field_names') ? '' : 'message')
-    ..aOS(4, const $core.bool.fromEnvironment('protobuf.omit_field_names') ? '' : 'provider')
-    ..aOS(5, const $core.bool.fromEnvironment('protobuf.omit_field_names') ? '' : 'createdAt')
-    ..aOS(6, const $core.bool.fromEnvironment('protobuf.omit_field_names') ? '' : 'expdate')
-    ..hasRequiredFields = false
-  ;
-
-  InboxMessage._() : super();
-  factory InboxMessage({
-    $core.String? uuid,
-    $core.int? status,
-    $core.String? message,
-    $core.String? provider,
-    $core.String? createdAt,
-    $core.String? expdate,
-  }) {
-    final _result = create();
-    if (uuid != null) {
-      _result.uuid = uuid;
-    }
-    if (status != null) {
-      _result.status = status;
-    }
-    if (message != null) {
-      _result.message = message;
-    }
-    if (provider != null) {
-      _result.provider = provider;
-    }
-    if (createdAt != null) {
-      _result.createdAt = createdAt;
-    }
-    if (expdate != null) {
-      _result.expdate = expdate;
-    }
-    return _result;
-  }
-  factory InboxMessage.fromBuffer($core.List<$core.int> i, [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) => create()..mergeFromBuffer(i, r);
-  factory InboxMessage.fromJson($core.String json) => InboxMessage.create()..mergeFromProto3Json($convert.jsonDecode(json));
-  @$core.Deprecated(
-  'Using this can add significant overhead to your binary. '
-  'Use [GeneratedMessageGenericExtensions.deepCopy] instead. '
-  'Will be removed in next major version')
-  InboxMessage clone() => InboxMessage()..mergeFromMessage(this);
-  @$core.Deprecated(
-  'Using this can add significant overhead to your binary. '
-  'Use [GeneratedMessageGenericExtensions.rebuild] instead. '
-  'Will be removed in next major version')
-  InboxMessage copyWith(void Function(InboxMessage) updates) => super.copyWith((message) => updates(message as InboxMessage)) as InboxMessage; // ignore: deprecated_member_use
-  $pb.BuilderInfo get info_ => _i;
-  @$core.pragma('dart2js:noInline')
-  static InboxMessage create() => InboxMessage._();
-  InboxMessage createEmptyInstance() => create();
-  static $pb.PbList<InboxMessage> createRepeated() => $pb.PbList<InboxMessage>();
-  @$core.pragma('dart2js:noInline')
-  static InboxMessage getDefault() => _defaultInstance ??= $pb.GeneratedMessage.$_defaultFor<InboxMessage>(create);
-  static InboxMessage? _defaultInstance;
-
-  @$pb.TagNumber(1)
-  $core.String get uuid => $_getSZ(0);
-  @$pb.TagNumber(1)
-  set uuid($core.String v) { $_setString(0, v); }
-  @$pb.TagNumber(1)
-  $core.bool hasUuid() => $_has(0);
-  @$pb.TagNumber(1)
-  void clearUuid() => clearField(1);
-
-  @$pb.TagNumber(2)
-  $core.int get status => $_getIZ(1);
-  @$pb.TagNumber(2)
-  set status($core.int v) { $_setSignedInt32(1, v); }
-  @$pb.TagNumber(2)
-  $core.bool hasStatus() => $_has(1);
-  @$pb.TagNumber(2)
-  void clearStatus() => clearField(2);
-
-  @$pb.TagNumber(3)
-  $core.String get message => $_getSZ(2);
-  @$pb.TagNumber(3)
-  set message($core.String v) { $_setString(2, v); }
-  @$pb.TagNumber(3)
-  $core.bool hasMessage() => $_has(2);
-  @$pb.TagNumber(3)
-  void clearMessage() => clearField(3);
-
-  @$pb.TagNumber(4)
-  $core.String get provider => $_getSZ(3);
-  @$pb.TagNumber(4)
-  set provider($core.String v) { $_setString(3, v); }
-  @$pb.TagNumber(4)
-  $core.bool hasProvider() => $_has(3);
-  @$pb.TagNumber(4)
-  void clearProvider() => clearField(4);
-
-  @$pb.TagNumber(5)
-  $core.String get createdAt => $_getSZ(4);
-  @$pb.TagNumber(5)
-  set createdAt($core.String v) { $_setString(4, v); }
-  @$pb.TagNumber(5)
-  $core.bool hasCreatedAt() => $_has(4);
-  @$pb.TagNumber(5)
-  void clearCreatedAt() => clearField(5);
-
-  @$pb.TagNumber(6)
-  $core.String get expdate => $_getSZ(5);
-  @$pb.TagNumber(6)
-  set expdate($core.String v) { $_setString(5, v); }
-  @$pb.TagNumber(6)
-  $core.bool hasExpdate() => $_has(5);
-  @$pb.TagNumber(6)
-  void clearExpdate() => clearField(6);
-}
-
 class Next extends $pb.GeneratedMessage {
   static final $pb.BuilderInfo _i = $pb.BuilderInfo(const $core.bool.fromEnvironment('protobuf.omit_message_names') ? '' : 'Next', package: const $pb.PackageName(const $core.bool.fromEnvironment('protobuf.omit_message_names') ? '' : 'vascularinbox.inbox'), createEmptyInstance: create)
     ..aOS(1, const $core.bool.fromEnvironment('protobuf.omit_field_names') ? '' : 'createdAt')
@@ -477,7 +371,7 @@ class Next extends $pb.GeneratedMessage {
     return _result;
   }
   factory Next.fromBuffer($core.List<$core.int> i, [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) => create()..mergeFromBuffer(i, r);
-  factory Next.fromJson($core.String json) => Next.create()..mergeFromProto3Json($convert.jsonDecode(json));
+  factory Next.fromJson($core.String i, [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) => create()..mergeFromJson(i, r);
   @$core.Deprecated(
   'Using this can add significant overhead to your binary. '
   'Use [GeneratedMessageGenericExtensions.deepCopy] instead. '